backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==7.8.0
twilio==8.10.3
sendgrid==6.11.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.17.0
google-cloud-sql-connector==1.4.3
firebase-admin==6.4.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
jinja2==3.1.2
celery==5.3.4
redis==5.0.1